+++ test/concurrent_test.dart
@@ -1,11 +1,12 @@
+// SPDX-FileCopyrightText: (c) 2021 Artsiom iG <github.com/rtmigo>
+// SPDX-License-Identifier: MIT
+
 import 'dart:async';
 import 'dart:math';
 
 import 'package:schedulers/src/b_base.dart';
 import 'package:schedulers/src/c_concurrent.dart';
 import 'package:test/test.dart';
-
-//import '../bin/source/pool.dart';
 
 class OmgError extends StateError {
   OmgError(final int x) : super(x.toString());
@@ -18,7 +19,7 @@ void main() {
 
   test("one", () async {
     final r = Random();
-    final pool = ParallelScheduler(concurrency: 4);
+    final pool = ParallelScheduler(4);
     final futures = List<Task<int>>.empty(growable: true);
     int maxEver = 0;
     for (int i = 0; i < 100; ++i) {
@@ -37,7 +38,9 @@ void main() {
         return 3;
       });
       futures.add(t);
-      //unawaited(t.result.whenComplete(() => 0));
+
+      // we want to handle OmgError errors, and the only way to do this,
+      // is to await results and catch the errors
       unawaited(Future.microtask(() async {
         try {
           await t.result;
@@ -57,8 +60,7 @@ void main() {
       }
     }
 
-    //await Future.delayed(Duration(seconds: 5));
-
+    expect(maxEver, 4);
     expect(errors, 25);
     expect(success, 75);
     expect(pool.currentlyRunning, 0);
@@ -68,7 +70,7 @@ void main() {
     test("exception from block is thrown to the zone", () async {
       bool gotError = false;
       await runZonedGuarded(() async {
-        final pool = ParallelScheduler(concurrency: 4);
+        final pool = ParallelScheduler(4);
         pool.run(() => throw "Oops!");
         await Future<void>.delayed(Duration(milliseconds: 100));
       }, (final _, final __) => gotError = true);
@@ -78,7 +80,7 @@ void main() {
     test("the same without exception", () async {
       bool gotError = false;
       await runZonedGuarded(() async {
-        final pool = ParallelScheduler(concurrency: 4);
+        final pool = ParallelScheduler(4);
         pool.run(() => 1);
         await Future<void>.delayed(Duration(milliseconds: 100));
       }, (final _, final __) => gotError = true);
@@ -88,7 +90,7 @@ void main() {
 
   test("million tasks", () async {
     // test whether too many tasks can lead to stack overflow
-    final pool = ParallelScheduler(concurrency: 32);
+    final pool = ParallelScheduler(32);
     final futures = List<Future<int>>.empty(growable: true);
     for (int i = 0; i < 1000000; ++i) {
       futures.add(pool.run(() async => i).result);