+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-- Result of `GetterFunc` changed to `FutureOr`
+# 1.0.0 
+
+- Added `ParallelScheduler`
+- Result of `GetterFunc` changed from `<R>` to `FutureOr<R>`
+- Results of tasks are now `await`ed by the schedulers
+- Dart version requirement increased to 2.18
 
 # 0.3.1